# Ignore max_compaction_bytes for compaction input that are within output key-range 