@rockeet 2024-03-24
`ignore_max_compaction_bytes_for_input` has a risk that makes very large compactions which makes large space amp, should limit the hard `max_compaction_bytes`(such as 2x of soft `max_compaction_bytes`) in this case?